+++ deepcave/plugins/static.py
@@ -158,6 +158,7 @@ class StaticPlugin(Plugin, ABC):
             raw_outputs_available = True
             for run in runs:
                 raw_outputs[run.id] = rc.get(run, self.id, inputs_key)
+
                 if raw_outputs[run.id] is None:
                     raw_outputs_available = False
 
@@ -171,6 +172,7 @@ class StaticPlugin(Plugin, ABC):
 
                     # Save for modal
                     self.raw_outputs = raw_outputs
+
                     outputs = self._process_raw_outputs(inputs, raw_outputs)
                     self._refresh_required = False
 
@@ -228,7 +230,6 @@ class StaticPlugin(Plugin, ABC):
 
                             # Save results in cache
                             # Same optional string problem
-                            # TODO: REMOVE WHEN DONE
 
                             rc.set(run, job_plugin_id, job_inputs_key, job_run_outputs)
                             self.logger.debug(f"Job {job_id} cached.")